Move the Pests class to a separate file. Document it.
